Moved theme.js from constants to utils
